Refactor type Lang.id to Integer

+++ src/main/java/io/github/mat3e/Lang.java
@@ -1,17 +1,17 @@
 package io.github.mat3e;
 
 class Lang {
-    private Long id;
+    private Integer id;
     private String welcameMsg;
     private String code;
 
-    public Lang(Long id, String welcameMsg, String code) {
+    public Lang(Integer id, String welcameMsg, String code) {
         this.id = id;
         this.welcameMsg = welcameMsg;
         this.code = code;
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
